requirements.txt
google-generativeai
chromadb
langchain-community
sentence-transformers
